requirements.txt
streamlit
langchain
langchain-groq
langchain-community
faiss-cpu
ollama
openai
python-dotenv
langchain-openai